requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6

# Image processing
Pillow>=10.1.0

# Excel handling
openpyxl>=3.1.2
pandas>=2.1.3

# PDF generation
reportlab>=4.0.7

# API and file handling
requests>=2.31.0
gdown>=4.7.1

# Utility packages
python-dotenv>=1.0.0
pydantic>=2.5.1